+++ src/DetailsView/reports/components/report-sections/minimal-rule-header.tsx
